Milestone 4 Deployment and Monitoring/Docker/requirements.txt
torch
torchvision
Pillow
# matplotlib  # Only needed for plotting in notebooks/scripts, not Flask app
numpy
Flask
